pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mapae"
version = "0.1.0"
description = "Mobile Authentication Platform via Automated Email"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.110.0",
  "uvicorn==0.27.1",
  "redis==5.0.1",
  "aiosmtpd==1.4.5",
  "python-multipart==0.0.9",
  "pyspf==2.0.14",
  "dnspython==2.6.1",
  "pydantic-settings==2.1.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["app*"]
